+++ packages/developer-portal/src/components/form-wizard/form-wizard-step.tsx
@@ -1,0 +1,52 @@
+import { Button, ButtonGroup } from '@reapit/elements'
+import React from 'react'
+import { FormWizardStepPropsInterface } from './interface'
+import { FieldValues, useForm } from 'react-hook-form'
+
+export const FormWizardStep = <TFieldValues extends FieldValues>({
+  formOptions,
+  component: Component,
+  previousStep,
+  canGoBack,
+  isLastStep,
+  onSubmit,
+  isSubmitting = false,
+}: FormWizardStepPropsInterface<TFieldValues>) => {
+  const {
+    handleSubmit,
+    register,
+    getValues,
+    watch,
+    setError,
+    setValue,
+    formState: { errors },
+  } = useForm(formOptions)
+
+  return (
+    <form style={{ marginTop: '1rem' }} onSubmit={handleSubmit(onSubmit)}>
+      <Component
+        register={register}
+        errors={errors}
+        getValues={getValues}
+        watch={watch}
+        setError={setError}
+        setValue={setValue}
+      />
+      <ButtonGroup>
+        <Button
+          disabled={!canGoBack}
+          onClick={(event) => {
+            event.preventDefault()
+            previousStep()
+          }}
+          intent="neutral"
+        >
+          Previous
+        </Button>
+        <Button disabled={isSubmitting} loading={isSubmitting} intent="primary">
+          {isLastStep ? 'Submit' : 'Next'}
+        </Button>
+      </ButtonGroup>
+    </form>
+  )
+}
