+++ packages/developer-portal/src/components/form-wizard/interface.ts
@@ -8,6 +8,7 @@ import {
   UseFormWatch,
   UseFormSetError,
   UseFormProps,
+  SubmitHandler,
 } from 'react-hook-form'
 
 export type StepComponent<TFieldValues extends FieldValues> = FC<{
@@ -20,6 +21,24 @@ export type StepComponent<TFieldValues extends FieldValues> = FC<{
 }>
 
 export type Step<TFieldValues extends FieldValues> = {
+  name: string
+  component: StepComponent<TFieldValues>
+  formOptions?: UseFormProps<TFieldValues>
+}
+
+export interface FormWizardPropsInterface<TFieldValues extends FieldValues> {
+  steps: Record<string, Step<TFieldValues>>
+  onSubmit: SubmitHandler<TFieldValues>
+  showStepNumbers?: boolean
+  isSubmitting?: boolean
+}
+
+export interface FormWizardStepPropsInterface<TFieldValues extends FieldValues> {
+  onSubmit: SubmitHandler<TFieldValues>
+  previousStep: () => void
+  canGoBack: boolean
+  isLastStep: boolean
+  isSubmitting?: boolean
   name: string
   component: StepComponent<TFieldValues>
   formOptions?: UseFormProps<TFieldValues>
